+++ HomeWork/Module7/003Makarenko/src/main/java/helper/Helper.java
@@ -7,11 +7,11 @@ public abstract class Helper<E> {
     protected SessionFactory sessionFactory = HibernateUtil.getSessionFactory();
     protected Session session;
 
-    public abstract E getById(long id);
+    public abstract E getById(int id);
 
     public abstract void add(E element);
 
-    public abstract void update(long id, E element);
+    public abstract void update(int id, E element);
 
     public SessionFactory getSessionFactory() {
         return sessionFactory;
